+++ Common/MVKCommonEnvironment.h
@@ -62,6 +62,14 @@ extern "C" {
 #	define MVK_TVOS					TARGET_OS_TV
 #endif
 
+/** Building for visionOS. */
+#ifndef TARGET_OS_XR
+#	define TARGET_OS_XR				0		// Older SDK's don't define TARGET_OS_XR
+#endif
+#ifndef MVK_VISIONOS
+#	define MVK_VISIONOS				TARGET_OS_XR
+#endif
+
 /** Building for iOS or tvOS. */
 #ifndef MVK_IOS_OR_TVOS
 #	define MVK_IOS_OR_TVOS			(MVK_IOS || MVK_TVOS)
@@ -70,6 +78,11 @@ extern "C" {
 /** Building for macOS or iOS. */
 #ifndef MVK_MACOS_OR_IOS
 #	define MVK_MACOS_OR_IOS			(MVK_MACOS || MVK_IOS)
+#endif
+
+/** Building for macOS, iOS or visionOS. */
+#ifndef MVK_MACOS_OR_IOS_OR_VISIONOS
+#	define MVK_MACOS_OR_IOS_OR_VISIONOS	(MVK_MACOS || MVK_IOS | MVK_VISIONOS)
 #endif
 
 /** Building for a Simulator. */